Fix lifetime duplication generating the same lifetime twice
Previously, the check for lifetime-collision only tested whether a
user-declared lifetime would conflict with the newly generated one.
However, it needs to also take into account the lifetimes it generated
in previous iterations. Consider the following struct

struct Test<'a, 'a_> {}

The old algorithm tried to duplicate 'a, saw that 'a_ already existing,
and called the duplicate 'a__. However, when it tried to duplicate 'a_,
it only looked at the user-declared lifetimes 'a and 'a_, and picked
'a__ again. This generated invalid code.

Now it takes into account previously-duped lifetimes, and would generate
'a___ in the example above.